# Codigo_Proyecto/Docker/electromovil-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de ElectroElite",
        "description": "API construida con Laravel y documentada con Swagger",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@tuservicio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/servicios": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Listar servicios según el rol",
                "operationId": "5f75a79845dec2d89cd58caedd04da06",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo_equipo",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de servicios"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear servicio",
                "operationId": "26cfc2cac3c017ab823806202789d470",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo_equipo",
                                    "marca",
                                    "modelo",
                                    "descripcion_problema"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer"
                                    },
                                    "tipo_equipo": {
                                        "type": "string"
                                    },
                                    "marca": {
                                        "type": "string"
                                    },
                                    "modelo": {
                                        "type": "string"
                                    },
                                    "descripcion_problema": {
                                        "type": "string"
                                    },
                                    "tecnico_id": {
                                        "type": "integer"
                                    },
                                    "costo": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servicios/{id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Mostrar un servicio",
                "operationId": "8e1ac7966c0ec20a4852a551a6a70352",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio encontrado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Actualizar servicio",
                "operationId": "34b60942afda348a33d50d09e2314ab7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servicio actualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Eliminar servicio",
                "operationId": "08260742e1e3efac8ef1abd9e4ef9ff1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Servicio eliminado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servicios/{id}/asignar-tecnico": {
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Asignar técnico a servicio",
                "operationId": "918b715a2afc172861f9880dbf3ab75f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tecnico_id"
                                ],
                                "properties": {
                                    "tecnico_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Técnico asignado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servicios/tecnico/mis-servicios": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Ver servicios asignados a un técnico",
                "operationId": "a5a89e546a8db38b84e5783652990622",
                "responses": {
                    "200": {
                        "description": "Lista de servicios asignados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}